src/test/javascript/spec/app/entities/professional/professional-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { Observable, of } from 'rxjs';

import { EdmsTestModule } from '../../../test.module';
import { ProfessionalUpdateComponent } from 'app/entities/professional/professional-update.component';
import { ProfessionalService } from 'app/entities/professional/professional.service';
import { Professional } from 'app/shared/model/professional.model';

describe('Component Tests', () => {
  describe('Professional Management Update Component', () => {
    let comp: ProfessionalUpdateComponent;
    let fixture: ComponentFixture<ProfessionalUpdateComponent>;
    let service: ProfessionalService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [EdmsTestModule],
        declarations: [ProfessionalUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(ProfessionalUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(ProfessionalUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(ProfessionalService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Professional(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Professional();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
